+++ packages/docs/docs/main/change-log.md
@@ -17,9 +17,16 @@ This page lists the changes to the IsaacScript framework.
   - `isDeadEnd`
   - `getRoomShapeNeighborGridIndexes`
   - `getNewRoomCandidatesBesideRoom`
+  - `getNewRoomCandidatesForLevel`
+  - `getNewRoomCandidate`
   - `hasCurse`
+  - `newRoom`
+  - `setRoomData`
+  - `getRoomDataForTypeVariant`
 - Changed the following helper functions:
   - `spawnCustomGridEntity` now supports the `breakable` argument, which will trigger `POST_GRID_ENTITY_CUSTOM_BROKEN` if it is hit by any type of explosion.
+- Added the following constants:
+  - `ALL_DISPLAY_FLAGS`
 
 ## July 27th, 2022